+++ routers/router_auth.py
@@ -5,6 +5,6 @@ from schemas.schema_auth import TokenResponse
 from schemas.schema_responses import SuccessResponse
 from controllers import controller_auth
 
-router = APIRouter(prefix="/auth", tags=["Auth"])
+router = APIRouter(prefix="/auth", tags=["Autenticación"])
 
 router.post("/login", response_model=SuccessResponse[TokenResponse])(controller_auth.login)