+++ routers/router_usuarios.py
@@ -5,7 +5,7 @@ from schemas.schema_responses import SuccessResponse
 from schemas.schema_usuarios import UsuarioAdminCreate, UsuarioUpdate, UsuarioResponse
 from controllers import controller_usuarios
 
-router = APIRouter(prefix="/admin/usuarios", tags=["Usuarios"])
+router = APIRouter(prefix="/admin/usuarios", tags=["Admin: Usuarios"])
 
 router.post("",               response_model=SuccessResponse[UsuarioResponse],        status_code=status.HTTP_201_CREATED)(controller_usuarios.crear_usuario)
 router.get("",                response_model=SuccessResponse[list[UsuarioResponse]]                                      )(controller_usuarios.listar_usuarios)